printf format/argument agreement in games/phantasia/misc.c.

+++ games/phantasia/misc.c
@@ -426,7 +426,7 @@ allstatslist()
 	mvprintw(13, 0, "Sin       : %9.5f", Player.p_sin);
 	mvprintw(14, 0, "Poison    : %9.5f", Player.p_poison);
 	mvprintw(15, 0, "Gems      : %9.0f", Player.p_gems);
-	mvprintw(16, 0, "Age       : %9d", Player.p_age);
+	mvprintw(16, 0, "Age       : %9d", (int)Player.p_age);
 	mvprintw(10, 40, "Holy Water: %9d", Player.p_holywater);
 	mvprintw(11, 40, "Amulets   : %9d", Player.p_amulets);
 	mvprintw(12, 40, "Charms    : %9d", Player.p_charms);
